pyproject.toml
[project]
name = "paragrafen-ai"
version = "0.1.0"
description = "Gratis svensk juridisk AI-assistent för allmänheten"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }

dependencies = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.7.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "feedparser>=6.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"
